+++ game.js
@@ -6,403 +6,443 @@ var gameSpeed = 1;
 
 // ==PROPER NAME==
 var properName = {
-  "dolans" : "Dolans",
-  "ambition" : "Ambition",
-  "experience" : "Experience",
-  "oil" : "Oil",
-  "dna" : "DNA Fragments",
-  "metal" : "Metal",
-  "data" : "Data"
-}
+  'dolans' : 'Dolans',
+  'ambition' : 'Ambition',
+  'experience' : 'Experience',
+  'oil' : 'Oil',
+  'dna' : 'DNA Fragments',
+  'metal' : 'Metal',
+  'data' : 'Data'
+};
 
 // ==CURRENCY==
 var currency = {
-  "dolans" : {
-    "amount" : 0,
-    "apc" : 1,
-    "aps" : 0,
-    "mult" : 1,
-    "max" : 1000
+  'dolans' : {
+    'amount' : 0,
+    'apc' : 1,
+    'aps' : 0,
+    'mult' : 1,
+    'max' : 1000
   },
-  "ambition" : {
-    "amount" : 0,
-    "apc" : 1,
-    "aps" : 0,
-    "mult" : 1,
-    "max" : 1000
+  'ambition' : {
+    'amount' : 0,
+    'apc' : 1,
+    'aps' : 0,
+    'mult' : 1,
+    'max' : 1000
   },
-  "experience" : {
-    "amount" : 0,
-    "apc" : 0,
-    "aps" : 0,
-    "mult" : 1,
-    "max" : 1000
+  'experience' : {
+    'amount' : 0,
+    'apc' : 0,
+    'aps' : 0,
+    'mult' : 1,
+    'max' : 1000
   },
-  "metal" : {
-    "amount" : 0,
-    "apc" : 0,
-    "aps" : 0,
-    "mult" : 1,
-    "max" : 1000
+  'metal' : {
+    'amount' : 0,
+    'apc' : 0,
+    'aps' : 0,
+    'mult' : 1,
+    'max' : 1000
   },
-  "oil" : {
-    "amount" : 0,
-    "apc" : 0,
-    "aps" : 0,
-    "mult" : 1,
-    "max" : 1000
+  'oil' : {
+    'amount' : 0,
+    'apc' : 0,
+    'aps' : 0,
+    'mult' : 1,
+    'max' : 1000
   },
-  "data" : {
-    "amount" : 0,
-    "apc" : 0,
-    "aps" : 0,
-    "mult" : 1,
-    "max" : 1000
+  'data' : {
+    'amount' : 0,
+    'apc' : 0,
+    'aps' : 0,
+    'mult' : 1,
+    'max' : 1000
   }
-}
+};
 
 // ==ROCKETS==
 var rockets = {
-  "slingshot" : {
-    "name" : "Slingshot",
-    "amount" : 0,
-    "tier" : 0,
-    "cost" : {
-      "ambition" : 10
+  'slingshot' : {
+    'name' : 'Slingshot',
+    'amount' : 0,
+    'tier' : 0,
+    'cost' : {
+      'ambition' : 10
     },
-    "aps" : {
-      "ambition" : 0.1
+    'aps' : {
+      'ambition' : 0.1
     }
   },
-  "bottlerocket" : {
-    "name" : "Bottle Rocket",
-    "amount" : 0,
-    "tier" : 0,
-    "cost" : {
-      "dolans" : 25
+  'bottlerocket' : {
+    'name' : 'Bottle Rocket',
+    'amount' : 0,
+    'tier' : 0,
+    'cost' : {
+      'dolans' : 25
     },
-    "aps" : {
-      "experience" : 0.1
+    'aps' : {
+      'experience' : 0.1
     }
   },
-  "pvcrocket" : {
-    "name" : "PVC Rocket",
-    "amount" : 0,
-    "tier" : 0,
-    "cost" : {
-      "dolans" : 100,
-      "ambition" : 25,
-      "experience" : 25
+  'pvcrocket' : {
+    'name' : 'PVC Rocket',
+    'amount' : 0,
+    'tier' : 0,
+    'cost' : {
+      'dolans' : 100,
+      'ambition' : 25,
+      'experience' : 25
     },
-    "aps" : {
-      "dolans" : 1,
-      "experience" : 1
+    'aps' : {
+      'dolans' : 1,
+      'experience' : 1
     }
   },
-  "compressedair" : {
-    "name" : "Compressed Air Rocket",
-    "amount" : 0,
-    "tier" : 1,
-    "cost" : {
-      "dolans" : 500,
-      "ambition" : 250,
-      "experience" : 150
+  'compressedair' : {
+    'name' : 'Compressed Air Rocket',
+    'amount' : 0,
+    'tier' : 1,
+    'cost' : {
+      'dolans' : 500,
+      'ambition' : 250,
+      'experience' : 150
     },
-    "aps" : {
-      "dolans" : 2,
-      "ambition" : 3,
-      "experience" : 3
+    'aps' : {
+      'dolans' : 2,
+      'ambition' : 3,
+      'experience' : 3
     }
   },
-  "rocketsim1" : {
-    "name" : "Rocket Simulator 1999",
-    "amount" : 0,
-    "tier" : 1,
-    "cost" : {
-      "dolans" : 1500,
-      "ambition" : 750,
-      "experience" : 1000
+  'rocketsim1' : {
+    'name' : 'Rocket Simulator 1999',
+    'amount' : 0,
+    'tier' : 1,
+    'cost' : {
+      'dolans' : 1500,
+      'ambition' : 750,
+      'experience' : 1000
     },
-    "aps" : {
-      "dolans" : 0,
-      "ambition" : 10,
-      "experience" : 10
+    'aps' : {
+      'dolans' : 0,
+      'ambition' : 10,
+      'experience' : 10
     }
   },
-  "rocketsim2" : {
-    "name" : "Rocket Simulator 2000",
-    "amount" : 0,
-    "tier" : 2,
-    "cost" : {
-      "dolans" : 3000,
-      "ambition" : 1500,
-      "experience" : 2000
+  'rocketsim2' : {
+    'name' : 'Rocket Simulator 2000',
+    'amount' : 0,
+    'tier' : 2,
+    'cost' : {
+      'dolans' : 3000,
+      'ambition' : 1500,
+      'experience' : 2000
     },
-    "aps" : {
-      "dolans" : 5,
-      "ambition" : 25,
-      "experience" : 25
+    'aps' : {
+      'dolans' : 5,
+      'ambition' : 25,
+      'experience' : 25
     }
   },
-  "spupnik1" : {
-    "name" : "Спупник-1",
-    "amount" : 0,
-    "tier" : 2,
-    "cost" : {
-      "dolans" : 3500,
-      "ambition" : 2500,
-      "metal" : 250
+  'spupnik1' : {
+    'name' : 'Спупник-1',
+    'amount' : 0,
+    'tier' : 2,
+    'cost' : {
+      'dolans' : 3500,
+      'ambition' : 2500,
+      'metal' : 250
     },
-    "aps" : {
-      "data" : 1,
-      "ambition" : 10,
-      "experience" : 35
+    'aps' : {
+      'data' : 1,
+      'ambition' : 10,
+      'experience' : 35
     }
   },
-  "luysin" : {
-    "name" : "Լուսին-4",
-    "amount" : 0,
-    "tier" : 2,
-    "cost" : {
-      "dolans" : 5000,
-      "ambition" : 2500,
-      "metal" : 250,
-      "oil" : 500
+  'luysin' : {
+    'name' : 'Լուսին-4',
+    'amount' : 0,
+    'tier' : 2,
+    'cost' : {
+      'dolans' : 5000,
+      'ambition' : 2500,
+      'metal' : 250,
+      'oil' : 500
     },
-    "aps" : {
-      "dolans" : 35,
-      "data" : 3
+    'aps' : {
+      'dolans' : 35,
+      'data' : 3
     }
   },
-  "m-aah" : {
-    "name" : "م اه",
-    "amount" : 0,
-    "tier" : 2,
-    "cost" : {
-      "dolans" : 10000,
-      "data" : 500,
-      "metal" : 1000,
-      "oil" : 2500
+  'm-aah' : {
+    'name' : 'م اه',
+    'amount' : 0,
+    'tier' : 2,
+    'cost' : {
+      'dolans' : 10000,
+      'data' : 500,
+      'metal' : 1000,
+      'oil' : 2500
     },
-    "aps" : {
-      "dolans" : 100,
-      "data" : 30
+    'aps' : {
+      'dolans' : 100,
+      'data' : 30
     }
   },
-}
+};
 
 // ==JOBS==
 var jobs = {
-  "mcdoodles" : {
-    "name" : "McDoodles",
-    "working" : false,
-    "active" : true,
-    "cost" : {
-      "experience" : 25
+  'mcdoodles' : {
+    'name' : 'McDoodles',
+    'working' : false,
+    'active' : true,
+    'cost' : {
+      'experience' : 25
     },
-    "apc" : {
-      "dolans" : 1
+    'apc' : {
+      'dolans' : 1
     }
   },
-  "stroopbucks" : {
-    "name" : "StoopBucks",
-    "working" : false,
-    "active" : true,
-    "cost" : {
-      "experience" : 100,
-      "ambition" : 500
+  'stroopbucks' : {
+    'name' : 'StoopBucks',
+    'working' : false,
+    'active' : true,
+    'cost' : {
+      'experience' : 100,
+      'ambition' : 500
     },
-    "apc" : {
-      "dolans" : 2
+    'apc' : {
+      'dolans' : 2
+    }
+  },
+  'itintern': {
+    'name': 'Macrohard Internship',
+    'working': false,
+    'active': true,
+    'cost': {
+      'experience': 5000,
+      'ambition': 1000
+    },
+    'apc': {
+      'dolans': 10,
+      'experience' : 2
+    }
+  },
+  'engineeringintern': {
+    'name': 'Engineering Intern',
+    'working': false,
+    'active': true,
+    'cost': {
+      'experience': 25000,
+      'ambition': 5000
+    },
+    'apc': {
+      'dolans': 50,
+      'experience' : 15
     }
   }
-}
+};
 
 // ==UPGRADES==
 var upgrades = {
-  "algebra" : {
-    "onetime" : 0,
-    "amount" : 0,
-    "prerequisites" : [],
-    "function" : function(){currency['dolans']['mult'] += .125; currency['experience']['mult'] += .125;},
-    "name" : "Learn Algebra",
-    "description" : "+ x1.125 multiplier to Dolans and Experience",
-    "cost" : {
-      "experience" : 1000
+  'algebra' : {
+    'onetime' : 0,
+    'amount' : 0,
+    'prerequisites' : [],
+    'function' : function(){currency['dolans']['mult'] += .125; currency['experience']['mult'] += .125;},
+    'name' : 'Learn Algebra',
+    'description' : '+ x1.125 multiplier to Dolans and Experience',
+    'cost' : {
+      'experience' : 1000
     }
   },
-  "geometry" : {
-    "onetime" : 0,
-    "amount" : 0,
-    "prerequisites" : ["algebra"],
-    "function" : function(){currency['dolans']['mult'] += .125; currency['experience']['mult'] += .125;},
-    "name" : "Learn Geometry",
-    "description" : "+ x1.125 multiplier to Dolans and Experience",
-    "cost" : {
-      "dolans" : 500,
-      "experience" : 1000
+  'geometry' : {
+    'onetime' : 0,
+    'amount' : 0,
+    'prerequisites' : ['algebra'],
+    'function' : function(){currency['dolans']['mult'] += .125; currency['experience']['mult'] += .125;},
+    'name' : 'Learn Geometry',
+    'description' : '+ x1.125 multiplier to Dolans and Experience',
+    'cost' : {
+      'dolans' : 500,
+      'experience' : 1000
     }
   },
-  "newBed" : {
-    "onetime" : 0,
-    "amount" : 0,
-    "prerequisites" : ["algebra"],
-    "function" : function(){currency['ambition']['mult'] += .25;},
-    "name" : "A new bed",
-    "description" : "+ x1.25 multiplier to Ambition",
-    "cost" : {
-      "dolans" : 1000
+  'newBed' : {
+    'onetime' : 0,
+    'amount' : 0,
+    'prerequisites' : ['algebra'],
+    'function' : function(){currency['ambition']['mult'] += .25;},
+    'name' : 'A new bed',
+    'description' : '+ x1.25 multiplier to Ambition',
+    'cost' : {
+      'dolans' : 1000
     }
   },
-  "moreRockets" : {
-    "onetime" : 0,
-    "amount" : 0,
-    "prerequisites" : ["geometry"],
-    "function" : function(){upgradeLevels['rocket']=1;for(rocket in rockets){writeRocketHTML(rocket);}},
-    "name" : "Better Rockets",
-    "description" : "Learn how to build more rockets",
-    "cost" : {
-      "dolans" : 1500,
-      "ambition" : 500,
-      "experience" : 1500
+  'moreRockets' : {
+    'onetime' : 0,
+    'amount' : 0,
+    'prerequisites' : ['geometry'],
+    'function' : function(){
+      upgradeLevels['rocket']=1;
+      for(rocket in rockets){
+        writeRocketHTML(rocket);
+      }
+    },
+    'name' : 'Better Rockets',
+    'description' : 'Learn how to build more rockets',
+    'cost' : {
+      'dolans' : 1500,
+      'ambition' : 500,
+      'experience' : 1500
     }
   },
-  "calculus" : {
-    "onetime" : 0,
-    "amount" : 0,
-    "prerequisites" : ["geometry"],
-    "function" : function(){currency['dolans']['mult'] += .25; currency['experience']['mult'] += .25;},
-    "name" : "Learn Calculus",
-    "description" : "+ x1.25 multiplier to Dolans and Experience",
-    "cost" : {
-      "dolans" : 5000,
-      "experience" : 1000
+  'calculus' : {
+    'onetime' : 0,
+    'amount' : 0,
+    'prerequisites' : ['geometry'],
+    'function' : function(){currency['dolans']['mult'] += .25; currency['experience']['mult'] += .25;},
+    'name' : 'Learn Calculus',
+    'description' : '+ x1.25 multiplier to Dolans and Experience',
+    'cost' : {
+      'dolans' : 5000,
+      'experience' : 1000
     }
   },
-  "moreRockets2" : {
-    "onetime" : 0,
-    "amount" : 0,
-    "prerequisites" : ["moreRockets", "calculus"],
-    "function" : function(){upgradeLevels['rocket']=2;for(r in rockets){writeRocketHTML(r);}upgradeLevels['building']=1;for(b in buildings){writeBuildingHTML(b);}},
-    "name" : "Even Better Rockets",
-    "description" : "Learn how to build more rockets",
-    "cost" : {
-      "dolans" : 7500,
-      "ambition" : 2500,
-      "experience" : 4500
+  'moreRockets2' : {
+    'onetime' : 0,
+    'amount' : 0,
+    'prerequisites' : ['moreRockets', 'calculus'],
+    'function' : function(){
+      upgradeLevels['rocket']=2;
+      for(r in rockets){
+        writeRocketHTML(r);
+      }
+      upgradeLevels['building']=1;
+      for(b in buildings){
+        writeBuildingHTML(b);
+      }
+    },
+    'name' : 'Even Better Rockets',
+    'description' : 'Learn how to build more rockets',
+    'cost' : {
+      'dolans' : 7500,
+      'ambition' : 2500,
+      'experience' : 4500
     }
   },
-}
+};
 
 // ==BUILDINGS==
 var buildings = {
-  "piggybank" : {
-    "name" : "Piggy Bank",
-    "description" : "Increase max Dolans by 1000",
-    "type" : "storage",
-    "amount" : 0,
-    "tier" : 0,
-    "function" : function(){ currency['dolans']['max']+=1000; },
-    "cost" : {
-      "dolans" : 500
+  'piggybank' : {
+    'name' : 'Piggy Bank',
+    'description' : 'Increase max Dolans by 1000',
+    'type' : 'storage',
+    'amount' : 0,
+    'tier' : 0,
+    'function' : function(){ currency['dolans']['max']+=1000; },
+    'cost' : {
+      'dolans' : 500
     }
   },
-  "capacityForGreatness" : {
-    "name" : "Capacity for greatness",
-    "description" : "Increase max Ambition by 1000",
-    "type" : "storage",
-    "amount" : 0,
-    "tier" : 0,
-    "function" : function(){ currency['ambition']['max']+=1000; },
-    "cost" : {
-      "ambition" : 500
+  'capacityForGreatness' : {
+    'name' : 'Capacity for greatness',
+    'description' : 'Increase max Ambition by 1000',
+    'type' : 'storage',
+    'amount' : 0,
+    'tier' : 0,
+    'function' : function(){ currency['ambition']['max']+=1000; },
+    'cost' : {
+      'ambition' : 500
     }
   },
-  "notebook" : {
-    "name" : "A pen and notebook",
-    "description" : "Increase max Experience by 1000",
-    "type" : "storage",
-    "amount" : 0,
-    "tier" : 0,
-    "function" : function(){ currency['experience']['max']+=1000; },
-    "cost" : {
-      "experience" : 500
+  'notebook' : {
+    'name' : 'A pen and notebook',
+    'description' : 'Increase max Experience by 1000',
+    'type' : 'storage',
+    'amount' : 0,
+    'tier' : 0,
+    'function' : function(){ currency['experience']['max']+=1000; },
+    'cost' : {
+      'experience' : 500
     }
   },
-  "metalstorage" : {
-    "name" : "Sheet Metal Storage",
-    "description" : "Increase max Metal by 1000",
-    "type" : "storage",
-    "amount" : 0,
-    "tier" : 1,
-    "function" : function(){ currency['metal']['max']+=1000; },
-    "cost" : {
-      "metal" : 500
+  'metalstorage' : {
+    'name' : 'Sheet Metal Storage',
+    'description' : 'Increase max Metal by 1000',
+    'type' : 'storage',
+    'amount' : 0,
+    'tier' : 1,
+    'function' : function(){ currency['metal']['max']+=1000; },
+    'cost' : {
+      'metal' : 500
     }
   },
-  "oiltank" : {
-    "name" : "Oil Tank",
-    "description" : "Increase max Oil by 1000",
-    "type" : "storage",
-    "amount" : 0,
-    "tier" : 1,
-    "function" : function(){ currency['oil']['max']+=1000; },
-    "cost" : {
-      "oil" : 500
+  'oiltank' : {
+    'name' : 'Oil Tank',
+    'description' : 'Increase max Oil by 1000',
+    'type' : 'storage',
+    'amount' : 0,
+    'tier' : 1,
+    'function' : function(){ currency['oil']['max']+=1000; },
+    'cost' : {
+      'oil' : 500
     }
   },
-  "mineMetal" : {
-    "name" : "Iron Mines",
-    "description" : "Increase Metal/second by 5",
-    "type" : "mine",
-    "amount" : 0,
-    "tier" : 1,
-    "function" : function(){ currency['metal']['aps']+=5; },
-    "cost" : {
-      "dolans" : 500
+  'mineMetal' : {
+    'name' : 'Iron Mines',
+    'description' : 'Increase Metal/second by 5',
+    'type' : 'mine',
+    'amount' : 0,
+    'tier' : 1,
+    'function' : function(){ currency['metal']['aps']+=5; },
+    'cost' : {
+      'dolans' : 500
     }
   },
-  "mineOil" : {
-    "name" : "Oil Wells",
-    "description" : "Increase Oil/second by 5",
-    "type" : "mine",
-    "amount" : 0,
-    "tier" : 1,
-    "function" : function(){ currency['oil']['aps']+=5; },
-    "cost" : {
-      "dolans" : 500
+  'mineOil' : {
+    'name' : 'Oil Wells',
+    'description' : 'Increase Oil/second by 5',
+    'type' : 'mine',
+    'amount' : 0,
+    'tier' : 1,
+    'function' : function(){ currency['oil']['aps']+=5; },
+    'cost' : {
+      'dolans' : 500
     }
   },
-}
+};
 
 //The upgrades you currently have
 var upgradesHave = [];
 
 //Upgrade levels
 var upgradeLevels = {
-  "rocket" : 0,
-  "building" : 0
-}
+  'rocket' : 0,
+  'building' : 0
+};
 
 //Worksafe?
 var worksafe = false;
 
 //Defaults
-var defaults = '"\"N4IgTg9gxg1gpgFwM4gFyiQGwJYDsDmSAFhAmqAIYC2EArrmagAwC+ANCAEakKZySxE5ENToM0rDgAcAblAHxGlGvUaSQUCFSlg4SJHAAmFbGGGjVE9uGiKk2KgEZzK8c2sLE9qgCYXYtWskKVopXGwYZ3QRV0DrUPwwCkM9c0x8OE4k/0t3DgytRDAATxy3dVw4AHcAISMywI4aXQAlWy8Gqw4oCkwoWkxaFGiLcutmuDbBZD8R2KtrTlpsTEM8QmEpbHx8Ys4KXBhOvI0KKQoobARigDEIMABxXQoESv1jitJMiAgjuYCuiAqIhekgEPcKBkPtYICsEAc/soAScqHg4ABZEGYaFNNEAeRW0OsACsIJxhqAqFBDD9DHwKSBcNQ4GgQOioAARWn0kAcKr3GDrNAAM1BcA4FwQ2BkLNQCDAtHFGggYOEcAAHlJ+Ng4LgoLKfABWaxnKDCGmYA7DRwsILyn5SJawBlM4GsgDK4IgUhqtGdvJA/LAgoIIrFEqgUplaHliu6KqUIA1WrAOr1sscTCYEqonCu2AguDQhqzJqkZuiFqtaB8tus/TAuj1pUrEEtuAZozUEvLaEcTQGjH7QIo6r7WfU1DzUsLHx7FeHVEHfaao/HpY4ye1uv1c5EveYA8wQ9XY9QmY3QKxe9NEiPJ5HZ4v6lh2P+uWz+4rn6Xx5Xj/XdRjHhG8Dx/Zdz1PQDbQ4BIkhSJAAAkKGjVAAG0AF1YKkRJkjgAAZOAZUwBlPG7LhllWIV3BYIAAA=\""'
+var defaults = 'N4IgTg9gxg1gpgFwM4gFyiQGwJYDsDmSAFhAmqAIYC2EArrmagAwC+ANCAEakKZySxE5ENToM0rDgAcAblAHxGlGvUaSQUCFSlg4SJHAAmFbGGGjVE9uGiKk2KgEZzK8c2sLE9qgCYXYtWskKVopXGwYZ3QRV0COTFoAT3tcf0t3DioAWgoKIjS3VmtQ/DAKQz1zTHw4TjKCuJAarUQwRIarDlw4AHcAISMOjJAaXQAlWy8h9SgKTChaBJRoi0LrUbgJwWQ/Fdira05abExDPEJhKWx8fETOClwYaetZqQoobAREgDEIMABxXQUBDdfTPLqkWoQCBPPYBTojRBzJAIP4UGrgkAQE4IB6w5Tw4ZUPBwACySMwmOJ3QA8idntYAFYQTjLUBUKCGaGGPhskA9P4wc5oABmyLgQQQkAgUiOsD5ArAQoIovF1k+eAQ/FS0UVyvwqswBmscAIJP45012uEeuFqDFRolL1oYF0uCg7XQxSkpXKegAEhQZHA0ABtAC6HBKZQqABk4MGjcJPGoOEcTmcVe4WEAAA=';
 
 //Create and save cookies
 function bake_cookie(name, value){
-	var exdate=new Date();
-	exdate.setDate(exdate.getDate() + 365);
-	var cookie = [name, '=', JSON.stringify(value),'; expires=.', exdate.toUTCString(), '; domain=.', window.location.host.toString(), '; path=/;'].join('');
-	document.cookie = cookie;
+  var exdate=new Date();
+  exdate.setDate(exdate.getDate() + 365);
+  var cookie = [name, '=', JSON.stringify(value),'; expires=.', exdate.toUTCString(), '; domain=.', window.location.host.toString(), '; path=/;'].join('');
+  document.cookie = cookie;
 }
 
 //Read cookie
 function eat_cookie(name){
-	var result = document.cookie.match(new RegExp(name + '=([^;]+)'));
-	result && (result = JSON.parse(result[1]));
-	return result;
+  var result = document.cookie.match(new RegExp(name + '=([^;]+)'));
+  result && (result = JSON.parse(result[1]));
+  return result;
 }
 
 //Make save data
@@ -421,9 +461,9 @@ function makeSaveData(){
   for (rocket in rockets){ saveData['rockets'][rocket] = {}; saveData['rockets'][rocket]['amount'] = rockets[rocket]['amount']; }
   for (upgrade in upgrades){ saveData['upgrades'][upgrade] = {}; saveData['upgrades'][upgrade]['amount'] = upgrades[upgrade]['amount']; }
   for (building in buildings){ saveData['buildings'][building] = {}; saveData['buildings'][building]['amount'] = buildings[building]['amount']; }
+  for (job in jobs) { saveData['jobs'][job] = {}; saveData['jobs'][job]['working'] = jobs[job]['working']; }
+  for (dolan in currency) { saveData['currency'][dolan] = currency['dolan']; }
   //Save all data, since all fields can be changed
-  saveData['jobs'] = jobs;
-  saveData['currency'] = currency;
   saveData['upgradesHave'] = upgradesHave;
   saveData['upgradeLevels'] = upgradeLevels;
   return saveData;
@@ -433,11 +473,11 @@ function makeSaveData(){
 function save(type){
   var saveData = makeSaveData();
   var saveData64 = LZString.compressToBase64(JSON.stringify(saveData));
-  if (type == "manual" || type == "auto"){
+  if (type == 'manual' || type == 'auto'){
     console.log('Saved game');
     bake_cookie('saveData', JSON.stringify(saveData64));
   }
-  else if (type == "export"){
+  else if (type == 'export'){
     $('#saveLoadTextarea').val(saveData64);
   }
 }
@@ -447,7 +487,7 @@ function load(type){
   //Set the save data to empty so no *major* errors occur
   var saveData = [];
   //Reset
-  if (type == "default"){
+  if (type == 'default'){
     try{
       saveData = JSON.parse(LZString.decompressFromBase64(defaults));
     }
@@ -456,7 +496,7 @@ function load(type){
     }
   }
   //Manual loading from import
-  if (type == "manual"){
+  if (type == 'manual'){
     try{
       saveData = JSON.parse(LZString.decompressFromBase64($('#saveLoadTextarea').val()));
     }
@@ -470,26 +510,26 @@ function load(type){
       saveData = JSON.parse(LZString.decompressFromBase64(eat_cookie('saveData')));
     }
     catch(e) {
-      console.log('Couldn\'t autoload')
+      console.log('Couldn\'t autoload');
     }
   }
   try{
     for (rocket in rockets){ rockets[rocket]['amount'] = saveData['rockets'][rocket]['amount']; }
     for (upgrade in upgrades){ upgrades[upgrade]['amount'] = saveData['upgrades'][upgrade]['amount']; }
     for (building in buildings){ buildings[building]['amount'] = saveData['buildings'][building]['amount']; }
-    jobs = saveData['jobs'];
-    currency = saveData['currency'];
+    for (job in jobs) { jobs[job]['working'] = saveData ['jobs'][job]['working']; }
+    for (dolan in currency) { currency[dolan] = saveData['currency'][dolan]; }
     upgradesHave = saveData['upgradesHave'];
     upgradeLevels = saveData['upgradeLevels'];
   }
   catch (e) {
-    console.log('An error occured while loading. Maybe there was no save file?')
+    console.log('An error occured while loading. Maybe there was no save file?');
   }
 }
 
 function createJob(job){
   var jobText = '';
-  jobText += '<tr class="spacer"></tr><tr id="area' + job + '"><td id="' + job + 'Button" onclick="getJob(\''+job+'\')" class="button greyed-out">Work at ' + jobs[job]['name'] + '</td><td >' + jobs[job]['name'] + ':</td><td class="amnt">'
+  jobText += '<tr class="spacer"></tr><tr id="area' + job + '"><td id="' + job + 'Button" onclick="getJob(\''+job+'\')" class="button greyed-out">Work at ' + jobs[job]['name'] + '</td><td >' + jobs[job]['name'] + ':</td><td class="amnt">';
   for (dolan in jobs[job]['apc']){
     jobText += '+' + jobs[job]['apc'][dolan] + ' ' + properName[dolan] + '/click; ';
   }
@@ -509,7 +549,7 @@ function writeJobHTML(job){
 //Creates HTML for buildings
 function createBuilding(building){
   if (buildings[building]['tier'] <= upgradeLevels['building'] && !$('#area'+building).length){
-    var buildingText = "";
+    var buildingText = '';
     buildingText += '<tr class="spacer"></tr><tr id="area' + building + '"><td id="' + building + 'Button" onclick="buildBuild(\''+building+'\')" class="button greyed-out">Build ' + buildings[building]['name'] + '</td>';
 
     buildingText += '<td id="' + building + 'Button10" onclick="buildBuild(\''+building+'\', 10)" class="smolbutton buildingBuy10 greyed-out">x10</td>';
@@ -532,7 +572,10 @@ function createBuilding(building){
 //Writes HTML for buildings
 function writeBuildingHTML(building){
   var rock = createBuilding(building);
-  if (rock != null){ $('#buildingTable' + buildings[building]['type'] + ' > tr:last').after(rock); }
+  if (rock != null) {
+    $('#buildingTable' + buildings[building]['type']).html($('#buildingTable' + buildings[building]['type']).html() + rock);
+  }
+  else {console.log('Building ' + building + ' is null');}
 }
 
 //Creates the HTML for upgrades
@@ -547,7 +590,7 @@ function createUpgrade(upgrade){
   if (neededUpgrades >= upgrades[upgrade]['prerequisites'].length && !$('#area'+upgrade).length){
     //Check if its a one time upgrade that's not been bought
     if ((upgrades[upgrade]['onetime'] == 0 && upgrades[upgrade]['amount'] == 0) || upgrades[upgrade]['onetime'] == 1){
-      var upgradeText = "";
+      var upgradeText = '';
       upgradeText += '<tr id="spacer' + upgrade + '" class="spacer"></tr><tr id="area' + upgrade + '"><td id="' + upgrade + 'Button" onclick="getUpgrade(\''+upgrade+'\')" class="button greyed-out">' + upgrades[upgrade]['name'] + '</td><td>' + upgrades[upgrade]['name'];
       if (upgrades[upgrade]['onetime'] == 1){
         upgradeText += ': </td><td class="amnt" id="amnt' + upgrade + '">' + upgrades[upgrade]['amount'] + '</td>"';
@@ -573,7 +616,7 @@ function writeUpgradeHTML(upgrade){
 //Creates the HTML for rockets
 function createRocket(rocket){
   if (rockets[rocket]['tier'] <= upgradeLevels['rocket'] && !$('#area'+rocket).length){
-    var rocketText = "";
+    var rocketText = '';
     rocketText += '<tr class="spacer"></tr><tr id="area' + rocket + '"><td id="' + rocket + 'Button" onclick="build(\''+rocket+'\')" class="button greyed-out">Build ' + rockets[rocket]['name'] + '</td>';
 
     rocketText += '<td id="' + rocket + 'Button10" onclick="build(\''+rocket+'\', 10)" class="smolbutton rocketBuy10 greyed-out">x10</td>';
@@ -677,7 +720,7 @@ function checkIfCanBuild(building, amount=1){
 function updateInfo(){
   for (dolan in currency){
     $('#' + dolan + 'Count').text(Math.floor(currency[dolan]['amount']));
-    $('#' + dolan + 'Max').text("(Max "+ Math.floor(currency[dolan]['max']) + ")");
+    $('#' + dolan + 'Max').text('(Max '+ Math.floor(currency[dolan]['max']) + ')');
   }
 }
 
@@ -696,7 +739,10 @@ function build(rocket, amount=1){
     for (dolan in rockets[rocket]['cost']){
       currency[dolan]['amount'] -= rockets[rocket]['cost'][dolan] * amount;
     }
-    $("#amnt"+rocket).text(rockets[rocket]['amount']);
+    for (dolan in rockets[rocket]['aps']){
+      currency[dolan]['aps'] += rockets[rocket]['aps'][dolan] * amount;
+    }
+    $('#amnt'+rocket).text(rockets[rocket]['amount']);
     updateInfo();
     greyOutRocket(rocket);
   }
@@ -709,7 +755,7 @@ function buildBuild(building, amount=1){
     for (dolan in buildings[building]['cost']){
       currency[dolan]['amount'] -= buildings[building]['cost'][dolan] * amount;
     }
-    $("#amnt"+building).text(buildings[building]['amount']);
+    $('#amnt'+building).text(buildings[building]['amount']);
     for (i = 0; i < amount; i++) { buildings[building]['function'](); }
     updateInfo();
     greyOutBuilding(building);
@@ -718,40 +764,40 @@ function buildBuild(building, amount=1){
 
 //Greys out rocket
 function greyOutRocket(rocket){
-  if (checkIfCanBuy(rocket)){$('#'+rocket+"Button").removeClass('greyed-out');}
-  else { $('#'+rocket+"Button").addClass('greyed-out'); }
-  if (checkIfCanBuy(rocket, 10)){$('#'+rocket+"Button10").removeClass('greyed-out');}
-  else { $('#'+rocket+"Button10").addClass('greyed-out'); }
-  if (checkIfCanBuy(rocket, 100)){$('#'+rocket+"Button100").removeClass('greyed-out');}
-  else { $('#'+rocket+"Button100").addClass('greyed-out'); }
-  if (checkIfCanBuy(rocket, 1000)){$('#'+rocket+"Button1000").removeClass('greyed-out');}
-  else { $('#'+rocket+"Button1000").addClass('greyed-out'); }
+  if (checkIfCanBuy(rocket)){$('#'+rocket+'Button').removeClass('greyed-out');}
+  else { $('#'+rocket+'Button').addClass('greyed-out'); }
+  if (checkIfCanBuy(rocket, 10)){$('#'+rocket+'Button10').removeClass('greyed-out');}
+  else { $('#'+rocket+'Button10').addClass('greyed-out'); }
+  if (checkIfCanBuy(rocket, 100)){$('#'+rocket+'Button100').removeClass('greyed-out');}
+  else { $('#'+rocket+'Button100').addClass('greyed-out'); }
+  if (checkIfCanBuy(rocket, 1000)){$('#'+rocket+'Button1000').removeClass('greyed-out');}
+  else { $('#'+rocket+'Button1000').addClass('greyed-out'); }
 }
 
 //Greys out building
 function greyOutBuilding(building){
-  if (checkIfCanBuild(building)){$('#'+building+"Button").removeClass('greyed-out');}
-  else { $('#'+building+"Button").addClass('greyed-out'); }
-  if (checkIfCanBuild(building, 10)){$('#'+building+"Button10").removeClass('greyed-out');}
-  else { $('#'+building+"Button10").addClass('greyed-out'); }
-  if (checkIfCanBuild(building, 100)){$('#'+building+"Button100").removeClass('greyed-out');}
-  else { $('#'+building+"Button100").addClass('greyed-out'); }
-  if (checkIfCanBuild(building, 1000)){$('#'+building+"Button1000").removeClass('greyed-out');}
-  else { $('#'+building+"Button1000").addClass('greyed-out'); }
+  if (checkIfCanBuild(building)){$('#'+building+'Button').removeClass('greyed-out');}
+  else { $('#'+building+'Button').addClass('greyed-out'); }
+  if (checkIfCanBuild(building, 10)){$('#'+building+'Button10').removeClass('greyed-out');}
+  else { $('#'+building+'Button10').addClass('greyed-out'); }
+  if (checkIfCanBuild(building, 100)){$('#'+building+'Button100').removeClass('greyed-out');}
+  else { $('#'+building+'Button100').addClass('greyed-out'); }
+  if (checkIfCanBuild(building, 1000)){$('#'+building+'Button1000').removeClass('greyed-out');}
+  else { $('#'+building+'Button1000').addClass('greyed-out'); }
 }
 
 //Adds currency when you click the buttons
 function work(type){
   currency[type]['amount'] += currency[type]['apc'] * currency[type]['mult'];
-  $("#"+type+"Count").text(Math.floor(currency[type]['amount']));
+  $('#'+type+'Count').text(Math.floor(currency[type]['amount']));
 }
 
 //Work a job
 function getJob(job){
   if (checkIfCanWork(job) && jobs[job]['working'] == false){
-    $("#"+job+"Button").removeClass('greyed-out');
-    $("#"+job+"Button").addClass('greyed-out-perm');
-    $("#"+job+"Button").text("You already work here");
+    $('#'+job+'Button').removeClass('greyed-out');
+    $('#'+job+'Button').addClass('greyed-out-perm');
+    $('#'+job+'Button').text('You already work here');
     jobs[job]['working'] = true;
     for (dolan in jobs[job]['apc']){
       currency[dolan]['apc'] += jobs[job]['apc'][dolan];
@@ -774,7 +820,7 @@ function getUpgrade(upgrade){
     }
     upgrades[upgrade]['amount']++;
     for (dolan in upgrades[upgrade]['cost']){
-      currency[dolan]['amount'] -= upgrades[upgrade]['cost'][dolan]
+      currency[dolan]['amount'] -= upgrades[upgrade]['cost'][dolan];
     }
     upgrades[upgrade]['function']();
   }
@@ -818,22 +864,22 @@ function initGame(){
   if (debug == false){ load('auto'); }
 
   //Jobs
-  var jobText = "";
+  var jobText = '';
   for (job in jobs){
     jobText += createJob(job);
   }
   $('#jobsTable').html(jobText);
 
   //Rockets
-  var rockitboi = "";
+  var rockitboi = '';
   for (rocket in rockets){
     rockitboi += createRocket(rocket);
   }
   $('#rocketsTable').html(rockitboi);
 
   //Buildings
-  var builditboi1 = "<h4>Capacity</h4>";
-  var builditboi2 = "<h4>Mines</h4><tr></tr>";
+  var builditboi1 = '<h4>Capacity</h4>';
+  var builditboi2 = '<h4>Mines</h4><tr></tr>';
   for (building in buildings){
     if (buildings[building]['type'] == 'storage')   { builditboi1 += createBuilding(building); }
     if (buildings[building]['type'] == 'mine')      { builditboi2 += createBuilding(building); }
@@ -844,7 +890,7 @@ function initGame(){
 
 
   //Upgrades
-  var upgradeText = "";
+  var upgradeText = '';
   for (upgrade in upgrades){
     upgradeText += createUpgrade(upgrade);
   }
@@ -863,20 +909,16 @@ function initGame(){
 function updateGame(){
   //Updates the multiplier
   for (dolan in currency){
-    if (currency[dolan]['mult'] > 1){
-      $('#' + dolan + 'Mult').text('(x' + currency[dolan]['mult'] + ')');
+    currency[dolan]['amount'] += currency[dolan]['aps'];
+    // if (currency[dolan]['mult'] > 1){
+    //   $('#' + dolan + 'Mult').text('(x' + currency[dolan]['mult'] + ')');
+    if (currency[dolan]['aps'] > 0){
+      $('#' + dolan + 'Mult').text(currency[dolan]['aps'] + '/sec');
     }
   }
 
   //Rocket loop
   for (rocket in rockets){
-    //Add currency
-    for (dolan in currency){
-      if (rockets[rocket]['aps'][dolan] != null){
-        currency[dolan]['amount'] += (rockets[rocket]['amount'] * rockets[rocket]['aps'][dolan] *  currency['dolans']['mult']) * gameSpeed;
-      }
-    }
-
     //Update rocket buy buttons
     greyOutRocket(rocket);
 
@@ -886,12 +928,12 @@ function updateGame(){
 
   //Update job buy buttons
   for (job in jobs){
-    if (checkIfCanWork(job)){$('#'+job+"Button").removeClass('greyed-out');}
+    if (checkIfCanWork(job)){$('#'+job+'Button').removeClass('greyed-out');}
   }
 
   //Update upgrade buy buttons
   for (upgrade in upgrades){
-    if (checkIfCanUpgrade(upgrade)){$('#'+upgrade+"Button").removeClass('greyed-out');}
+    if (checkIfCanUpgrade(upgrade)){$('#'+upgrade+'Button').removeClass('greyed-out');}
   }
 
   for (building in buildings){